[Feature #163020921] Refactored code errors
